Shopping cart
You enter a name, e.g., a product, into an input field. You have the option to add this product to an array. You can also reset the array to zero by pressing a button. Additionally, you can randomly select a product from the list of entered products and display the contents of the entire array.